Fixed several warnings

+++ src/bootloader/make_initrd.c
@@ -1,4 +1,6 @@
 #include <stdio.h>
+#include <string.h>
+#include <stdlib.h>
 
 struct initrd_header {
     unsigned char magic;
@@ -10,7 +12,7 @@ struct initrd_header {
 int main(char argc, char **argv) {
     int nheaders = (argc - 1) / 2;
     struct initrd_header headers[64];
-    printf("size of header: %d\n", sizeof(struct initrd_header));
+    printf("size of header: %d\n", (int) sizeof(struct initrd_header));
     unsigned int off = sizeof(struct initrd_header) * 64 + sizeof(int);
 
     for (int i = 0; i < nheaders; i++) {
